[DF-7045] fixed  warnings
Former-commit-id: 17fa9649da57ed5b0e23810e0986603334b54a48

+++ Sources/Internal/Platform/TemplateWin32/WinUAPXamlApp.cpp
@@ -96,7 +96,7 @@ void WinUAPXamlApp::SetScreenMode(ApplicationViewWindowingMode screenMode)
 
 Windows::Foundation::Size WinUAPXamlApp::GetCurrentScreenSize()
 {
-    return Windows::Foundation::Size(physicalWidth, physicalHeight);
+    return Windows::Foundation::Size(static_cast<float32>(physicalWidth), static_cast<float32>(physicalHeight));
 }
 
 void WinUAPXamlApp::SetCursorPinning(bool isPinning)
